Login: Move magic link into its own component
This also adds a new route for this component.
