+++ client/login/index.js
@@ -2,10 +2,14 @@
  * Internal dependencies
  */
 import config from 'config';
-import { login } from './controller';
+import { login, magicLogin } from './controller';
 import { makeLayout, redirectLoggedIn } from 'controller';
 
 export default router => {
+	if ( config.isEnabled( 'magic-login' ) ) {
+		router( '/login/link', redirectLoggedIn, magicLogin, makeLayout );
+	}
+
 	if ( config.isEnabled( 'wp-login' ) ) {
 		router( '/login/:twoFactorAuthType?', redirectLoggedIn, login, makeLayout );
 	}
