+++ client/login/wp-login/index.jsx
@@ -5,93 +5,35 @@ import React, { PropTypes } from 'react';
 import { compact } from 'lodash';
 import { connect } from 'react-redux';
 import { localize } from 'i18n-calypso';
+import page from 'page';
 
 /**
  * Internal dependencies
  */
-import {
-	CHECK_YOUR_EMAIL_PAGE,
-	INTERSTITIAL_PAGE,
-	LINK_EXPIRED_PAGE,
-	REQUEST_FORM,
-} from 'state/login/magic-login/constants';
 import config, { isEnabled } from 'config';
-import EmailedLoginLinkSuccessfully from '../magic-login/emailed-login-link-successfully';
-import EmailedLoginLinkExpired from '../magic-login/emailed-login-link-expired';
-import {
-	getMagicLoginEmailAddressFormInput,
-	getMagicLoginCurrentView,
-} from 'state/selectors';
 import { getCurrentQueryArguments } from 'state/ui/selectors';
 import Gridicon from 'gridicons';
-import HandleEmailedLinkForm from '../magic-login/handle-emailed-link-form';
-import {
-	hideMagicLoginRequestForm,
-	showMagicLoginInterstitialPage,
-	showMagicLoginRequestForm,
-} from 'state/login/magic-login/actions';
 import Main from 'components/main';
 import LoginBlock from 'blocks/login';
-import RequestLoginEmailForm from '../magic-login/request-login-email-form';
 import { recordTracksEvent } from 'state/analytics/actions';
 import PageViewTracker from 'lib/analytics/page-view-tracker';
 import GlobalNotices from 'components/global-notices';
 import notices from 'notices';
+import { login } from 'lib/paths';
 
 class Login extends React.Component {
 	static propTypes = {
-		hideMagicLoginRequestForm: PropTypes.func.isRequired,
-		magicLoginEmailAddress: PropTypes.string,
-		magicLoginEnabled: PropTypes.bool,
-		magicLoginView: PropTypes.string,
 		recordTracksEvent: PropTypes.func.isRequired,
-		showMagicLoginInterstitialPage: PropTypes.func.isRequired,
-		showMagicLoginRequestForm: PropTypes.func.isRequired,
 		translate: PropTypes.func.isRequired,
 		twoFactorAuthType: PropTypes.string,
-	};
-
-	onClickEnterPasswordInstead = event => {
-		event.preventDefault();
-		this.props.recordTracksEvent( 'calypso_login_enter_password_instead_click' );
-		this.props.hideMagicLoginRequestForm();
 	};
 
 	onMagicLoginRequestClick = event => {
 		event.preventDefault();
 		this.props.recordTracksEvent( 'calypso_login_magic_login_request_click' );
-		this.props.showMagicLoginRequestForm();
+
+		page( login( { twoFactorAuthType: 'link' } ) );
 	};
-
-	magicLoginMainContent() {
-		const {
-			magicLoginView,
-			magicLoginEmailAddress,
-		} = this.props;
-
-		switch ( magicLoginView ) {
-			case LINK_EXPIRED_PAGE:
-				this.props.recordTracksEvent( 'calypso_login_magic_link_expired_link_view' );
-				return <EmailedLoginLinkExpired />;
-			case CHECK_YOUR_EMAIL_PAGE:
-				this.props.recordTracksEvent( 'calypso_login_magic_link_link_sent_view' );
-				return <EmailedLoginLinkSuccessfully emailAddress={ magicLoginEmailAddress } />;
-			case INTERSTITIAL_PAGE:
-				this.props.recordTracksEvent( 'calypso_login_magic_link_interstitial_view' );
-				return <HandleEmailedLinkForm />;
-		}
-	}
-
-	componentWillMount() {
-		const {
-			magicLoginEnabled,
-			queryArguments,
-		} = this.props;
-
-		if ( magicLoginEnabled && queryArguments && queryArguments.action === 'handleLoginEmail' ) {
-			this.props.showMagicLoginInterstitialPage();
-		}
-	}
 
 	goBack = event => {
 		event.preventDefault();
@@ -104,32 +46,20 @@ class Login extends React.Component {
 	};
 
 	footerLinks() {
-		const {
-			magicLoginEnabled,
-			magicLoginView,
-			translate,
-		} = this.props;
+		const { translate } = this.props;
 
-		if ( magicLoginEnabled && magicLoginView === REQUEST_FORM ) {
-			return <a href="#"
-				key="enter-password-link"
-				onClick={ this.onClickEnterPasswordInstead }>
-					{ translate( 'Enter a password instead' ) }
-				</a>;
-		}
-
-		const goBackLink = ! magicLoginView && <a
+		const goBackLink = <a
 			href="#"
 			key="back-link"
 			onClick={ this.goBack }>
 				<Gridicon icon="arrow-left" size={ 18 } /> { this.props.translate( 'Return' ) }
 			</a>;
-		const showMagicLoginLink = magicLoginEnabled && ! magicLoginView && <a href="#"
+		const showMagicLoginLink = isEnabled( 'magic-login' ) && <a href="#"
 			key="magic-login-link"
 			onClick={ this.onMagicLoginRequestClick }>
 				{ translate( 'Email me a login link' ) }
 			</a>;
-		const resetPasswordLink = ! magicLoginView && <a
+		const resetPasswordLink = <a
 			href={ config( 'login_url' ) + '?action=lostpassword' }
 			key="lost-password-link">
 				{ this.props.translate( 'Lost your password?' ) }
@@ -144,7 +74,6 @@ class Login extends React.Component {
 
 	render() {
 		const {
-			magicLoginView,
 			queryArguments,
 			translate,
 			twoFactorAuthType,
@@ -156,42 +85,30 @@ class Login extends React.Component {
 
 				<GlobalNotices id="notices" notices={ notices.list } />
 
-				{ this.magicLoginMainContent() || (
-					<div>
-						<div className="wp-login__container">
-							{ magicLoginView === REQUEST_FORM
-								? <RequestLoginEmailForm />
-								: <LoginBlock
-									twoFactorAuthType={ twoFactorAuthType }
-									redirectLocation={ queryArguments.redirect_to }
-									title={ translate( 'Log in to your account.' ) } />
-							}
-						</div>
-						<div className="wp-login__footer">
-							{ this.footerLinks() }
-						</div>
+				<div>
+					<div className="wp-login__container">
+						<LoginBlock
+							twoFactorAuthType={ twoFactorAuthType }
+							redirectLocation={ queryArguments.redirect_to }
+							title={ translate( 'Log in to your account.' ) } />
 					</div>
-				) }
+					<div className="wp-login__footer">
+						{ this.footerLinks() }
+					</div>
+				</div>
 			</Main>
 		);
 	}
 }
 
 const mapState = state => {
-	const magicLoginEnabled = isEnabled( 'magic-login' );
 	return {
-		magicLoginEnabled,
-		magicLoginEmailAddress: getMagicLoginEmailAddressFormInput( state ),
-		magicLoginView: magicLoginEnabled ? getMagicLoginCurrentView( state ) : null,
 		queryArguments: getCurrentQueryArguments( state ),
 	};
 };
 
 const mapDispatch = {
-	hideMagicLoginRequestForm,
-	showMagicLoginInterstitialPage,
-	showMagicLoginRequestForm,
-	recordTracksEvent,
+	recordTracksEvent
 };
 
 export default connect( mapState, mapDispatch )( localize( Login ) );
